requirements.txt
streamlit
numpy
pillow
tensorflow
ultralytics
deep-translator
google-generativeai
transformers
torch
torchvision
nltk
tqdm